requirements.txt
numpy==1.24.2 
pandas==2.0.0
jinja2==3.1.2
openpyxl==3.0.10
avro==1.11.1
tqdm==4.65.0
impyla==0.18.0
python-dotenv==1.0.0
maskpass==0.3.7
pyperclip==1.8.2
sql_metadata==2.8.0
hdbcli==2.16.26
requests==2.31.0
fastapi==0.96.0
flatten_json==0.1.13
XlsxWriter==3.1.2
robotframework==6.1.1
fsspec==2023.9.0